update: add merkle.ts to sdk
