+++ common/sdk/src/index.ts
@@ -4,3 +4,4 @@ export { default as KyveWebClient } from "./clients/rpc-client/web.client";
 export * as constants from "./constants";
 export { registry } from "./registry/tx.registry";
 export { KyveSDK as default } from "./sdk";
+export { dataItemToSha256, generateMerkleRoot } from "./merkle/merkle"
