+++ integrations/tendermint/utils/merkle.ts
@@ -1,4 +1,5 @@
 import * as crypto from '@cosmjs/crypto';
+import { dataItemToSha256, generateMerkleRoot } from "@kyvejs/sdk";
 
 // Creates an Array of hashes from an array of data items (bundle).
 // The hash of a data item consists of the Merkle root from the block and the block
@@ -20,37 +21,4 @@ function tendermintDataItemToSha256(key: string, merkleRoot: Uint8Array): Uint8A
   const combined = Buffer.concat([keyBytes, merkleRoot]);
 
   return crypto.sha256(combined);
-}
-
-function dataItemToSha256(data: any): Uint8Array {
-  // Encode the serialized object to UTF-8
-  const encoded_obj: Uint8Array = Buffer.from(JSON.stringify(data), 'utf-8');
-  // Calculate the SHA-256 hash
-  return crypto.sha256(encoded_obj)
-}
-
-export function generateMerkleRoot(hashes: Uint8Array[]): Uint8Array {
-  if (!hashes || hashes.length == 0) {
-    return Buffer.from("");
-  }
-
-  // Ensure number of hashes (leafs) are even by copying the
-  // last hash (the very right leaf) if the amount is odd
-  if (hashes.length % 2 !== 0) {
-    hashes.push(hashes[hashes.length - 1]);
-  }
-
-  const combinedHashes: Uint8Array[] = [];
-  for(let i = 0; i < hashes.length; i += 2) {
-    const hashesConcatenated = new Uint8Array([ ...hashes[i], ...hashes[i+1]])
-    const hash = crypto.sha256(hashesConcatenated);
-    combinedHashes.push(hash);
-  }
-
-  // If the combinedHashes length is 1, it means that we have the merkle root already,
-  // and we can return the hex representation
-  if (combinedHashes.length === 1) {
-    return combinedHashes[0];
-  }
-  return generateMerkleRoot(combinedHashes);
 }
